requirements.txt
# requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
black==24.8.0
ruff==0.6.9
isort==5.13.2

# Additional for cash session integration
pydantic==2.5.0
